requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
flask-cors
flask-limiter
flask-wtf
itsdangerous
pyotp
redis
celery
anthropic
openai
weasyprint
python-dotenv
psycopg2-binary
pytest
requests
